feat(widget): create navbar logo and item
