+++ lib/widgets/navigation_bar/navigation_bar.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:the_basics/widgets/navigation_bar/navbar_item.dart';
+import 'package:the_basics/widgets/navigation_bar/navbar_logo.dart';
 
 class NavigationBar extends StatelessWidget {
   @override
@@ -8,39 +10,19 @@ class NavigationBar extends StatelessWidget {
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: <Widget>[
-          SizedBox(
-            height: 80,
-            width: 150,
-            child: Image.asset('assets/logo.png'),
-          ),
+          NavBarLogo(),
           Row(
             mainAxisSize: MainAxisSize.min,
             children: <Widget>[
-              _NavBarItem('Episodes'),
+              NavBarItem('Episodes'),
               SizedBox(
                 width: 60,
               ),
-              _NavBarItem('About'),
+              NavBarItem('About'),
             ],
           )
         ],
       ),
-    );
-  }
-}
-
-class _NavBarItem extends StatelessWidget {
-  final String title;
-  const _NavBarItem(
-    this.title, {
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      title,
-      style: TextStyle(fontSize: 18),
     );
   }
 }
